Project 3/requirements.txt
pipecat-ai[silero,deepgram,openai,elevenlabs,local]
python-dotenv
